expose render prop types, refactor default renderers
