+++ packages/slate-react/src/components/leaf.tsx
@@ -1,15 +1,10 @@
 import React from 'react'
 import { Text, Element } from 'slate'
+
 import String from './string'
-import {
-  CustomDecoration,
-  CustomDecorationProps,
-  CustomMark,
-  CustomMarkProps,
-  PlaceholderDecoration,
-} from './custom'
-import { PLACEHOLDER_SYMBOL } from '../utils/weak-maps'
 import { Leaf as SlateLeaf } from '../utils/leaf'
+import { PLACEHOLDER_SYMBOL } from '../utils/weak-maps'
+import { RenderDecorationProps, RenderMarkProps } from './editable'
 
 /**
  * Individual leaves in a text node with unique formatting.
@@ -19,8 +14,8 @@ const Leaf = (props: {
   isLast: boolean
   leaf: SlateLeaf
   parent: Element
-  renderDecoration?: (props: CustomDecorationProps) => JSX.Element
-  renderMark?: (props: CustomMarkProps) => JSX.Element
+  renderDecoration?: (props: RenderDecorationProps) => JSX.Element
+  renderMark?: (props: RenderMarkProps) => JSX.Element
   text: Text
 }) => {
   const {
@@ -28,10 +23,10 @@ const Leaf = (props: {
     isLast,
     text,
     parent,
-    renderDecoration = (props: CustomDecorationProps) => (
-      <CustomDecoration {...props} />
+    renderDecoration = (props: RenderDecorationProps) => (
+      <DefaultDecoration {...props} />
     ),
-    renderMark = (props: CustomMarkProps) => <CustomMark {...props} />,
+    renderMark = (props: RenderMarkProps) => <DefaultMark {...props} />,
   } = props
 
   let children = (
@@ -94,5 +89,51 @@ const MemoizedLeaf = React.memo(Leaf, (prev, next) => {
     SlateLeaf.equals(next.leaf, prev.leaf)
   )
 })
+
+/**
+ * The default custom decoration renderer.
+ */
+
+export const DefaultDecoration = (props: RenderDecorationProps) => {
+  const { attributes, children } = props
+  return <span {...attributes}>{children}</span>
+}
+
+/**
+ * The default custom mark renderer.
+ */
+
+export const DefaultMark = (props: RenderMarkProps) => {
+  const { attributes, children } = props
+  return <span {...attributes}>{children}</span>
+}
+
+/**
+ * A custom decoration for the default placeholder behavior.
+ */
+
+const PlaceholderDecoration = (props: RenderDecorationProps) => {
+  const { decoration, attributes, children } = props
+  const { placeholder } = decoration
+  return (
+    <span {...attributes}>
+      <span
+        contentEditable={false}
+        style={{
+          pointerEvents: 'none',
+          display: 'inline-block',
+          verticalAlign: 'text-top',
+          width: '0',
+          maxWidth: '100%',
+          whiteSpace: 'nowrap',
+          opacity: '0.333',
+        }}
+      >
+        {placeholder}
+      </span>
+      {children}
+    </span>
+  )
+}
 
 export default MemoizedLeaf
